+++ CHANGELOG.md
@@ -13,6 +13,12 @@
 - Change SchemeThemeListener, improve handle icon dark/light
 - Rename theme_dark_light -> color_dark_light
 - Add zukan_listener_enabled setting
+- Add zukan_reporter
+- Rename folder commands -> core
+- Rename settings -> zukan_pref_settings
+- Merge folder events -> core
+- Move move_folders to helpers
+- Move clean_comments to helpers
 
 ## [0.4.2] - 2024-12-19